Only suggest names that make sense (#19843)
Previously, when writing a data constructor in a pattern that doesn't
exist, GHC would suggest similar names, even if they aren't data
constructors or pattern synonyms. This patch fixes this, along with only
showing suggestions for record fields in record updates and record dot
notation, even when -XNoFieldSelectors is enabled, which wasn't the case
previously.

Fixes #19843.